split subject to shared and internal
